+++ internal/config/gitconfig/high_level.go
@@ -138,8 +138,8 @@ func RemovePerennialRegex(runner subshelldomain.Runner) error {
 	return RemoveConfigValue(runner, configdomain.ConfigScopeLocal, configdomain.KeyPerennialRegex)
 }
 
-func RemoveProposalsShowLineage(runner subshelldomain.Runner) error {
-	return RemoveConfigValue(runner, configdomain.ConfigScopeLocal, configdomain.KeyProposalsShowLineage)
+func RemoveProposalBreadcrumb(runner subshelldomain.Runner) error {
+	return RemoveConfigValue(runner, configdomain.ConfigScopeLocal, configdomain.KeyProposalBreadcrumb)
 }
 
 func RemoveProposalsShowLineageSingleStack(runner subshelldomain.Runner) error {
@@ -307,8 +307,8 @@ func SetPerennialRegex(runner subshelldomain.Runner, value configdomain.Perennia
 	return SetConfigValue(runner, scope, configdomain.KeyPerennialRegex, value.String())
 }
 
-func SetProposalsShowLineage(runner subshelldomain.Runner, value forgedomain.ProposalsShowLineage, scope configdomain.ConfigScope) error {
-	return SetConfigValue(runner, scope, configdomain.KeyProposalsShowLineage, value.String())
+func SetProposalBreadcrumb(runner subshelldomain.Runner, value forgedomain.ProposalBreadcrumb, scope configdomain.ConfigScope) error {
+	return SetConfigValue(runner, scope, configdomain.KeyProposalBreadcrumb, value.String())
 }
 
 func SetProposalsShowLineageSingleStack(runner subshelldomain.Runner, value forgedomain.ProposalsShowLineageSingleStack, scope configdomain.ConfigScope) error {